pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["IL-studio <welcome@il-studio.ru>"]
readme = "README.md"
packages = [{include = "example"}]

[tool.poetry.dependencies]
django = "^4.2"
django-braces = "^1.15.0"
django-cleanup = "^7.0.0"
django-extensions = "^3.2.3"
django-extra-views = "^0.14.0"
django-loginas = "^0.3.11"
django-model-utils = "^4.3.1"
django-nplusone = "^0.0.5"
django-querycount = "^0.8.3"
django-split-settings = "^1.2.0"
django-spurl = "^0.6.8"
django-stubs = "^1.15.0"
ipython = "^8.14.0"
pillow = "^10.0.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
python = "^3.11"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
singlemodeladmin = "^0.9"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"
psycopg2-binary = "^2.9.6"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = [
    "E",
    "F",
    "T20",
    "UP",
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".env",
    ".env_legacy",
    "migrations",
    "__init__.py",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.per-file-ignores]
"import/*" = ["T20"]
